+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Z-Ray-Doctrine2 CHANGELOG
 
+## 1.1.0 (2015-07-27)
+
+* All entity mappings now displayed
+* Added count of executed queries with cache info
+* Added parameters to queries
+* Added SQL transaction tracking
+* Added query syntax highlighting expand text support (show more / show less)
+* Added query number
+* Added cache tab
+
 ## Version 1.0.0 (2015-05-07)
 
 * Added entities tab